pyproject.toml
[project]
name = "project-big-data-2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "joblib>=1.5.1",
    "kafka-python>=2.2.10",
    "pandas>=2.2.3",
    "scikit-learn>=1.6.1",
]
